fix scanner construction
